requirements.txt
openai
google-auth
google-cloud-secret-manager
pillow